requirements.txt
langchain==0.3.7
langchain-openai==0.2.6
langchain-ollama==0.2.1
langgraph==0.2.53
openai==1.54.4
python-dotenv==1.0.1

